@@ -19,41 +19,40 @@
   },
   "homepage": "https://github.com/The-Politico/The-Politico.github.io#readme",
   "dependencies": {
-    "@fortawesome/fontawesome": "^1.1.4",
-    "@fortawesome/fontawesome-free-brands": "^5.0.8",
-    "@fortawesome/fontawesome-free-solid": "^5.0.8",
-    "@fortawesome/react-fontawesome": "^0.0.17",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
-    "bulma": "^0.6.2",
-    "normalize.css": "^7.0.0",
-    "parcel-plugin-react-markdown": "^0.0.1",
-    "politico-style": "git+ssh://git@github.com/The-Politico/politico-style",
-    "preact": "^8.2.7",
-    "preact-compat": "^3.17.0",
-    "react": "^16.2.0",
-    "react-addons-css-transition-group": "^15.6.2",
-    "react-dom": "^16.2.0",
-    "react-markdown": "^3.2.2",
-    "react-router-dom": "^4.2.2",
-    "react-transition-group": "^2.2.1"
+    "@fortawesome/fontawesome": "1.1.4",
+    "@fortawesome/fontawesome-free-brands": "5.0.8",
+    "@fortawesome/fontawesome-free-solid": "5.0.8",
+    "@fortawesome/react-fontawesome": "0.0.17",
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-stage-3": "6.24.1",
+    "bulma": "0.6.2",
+    "normalize.css": "7.0.0",
+    "parcel-plugin-react-markdown": "0.0.1",
+    "politico-style": "^1.0.1",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-markdown": "^3.3.4",
+    "react-router-dom": "^4.3.1",
+    "react-transition-group": "^2.4.0"
   },
   "devDependencies": {
-    "babel-plugin-module-resolver": "^3.0.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-preset-env": "^1.6.1",
+    "babel-plugin-module-resolver": "3.0.0",
+    "babel-plugin-transform-react-jsx": "6.24.1",
+    "babel-plugin-wildcard": "^5.0.0",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "eslint": "^4.16.0",
-    "eslint-config-standard": "^11.0.0-beta.0",
-    "eslint-config-standard-react": "^5.0.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.6.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "node-sass": "^4.7.2",
-    "nunjucks": "^3.0.1",
-    "parcel-bundler": "*",
-    "parcel-plugin-nunjucks": "git://github.com/The-Politico/parcel-plugin-nunjucks.git"
+    "eslint": "4.16.0",
+    "eslint-config-standard": "11.0.0-beta.0",
+    "eslint-config-standard-react": "5.0.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-node": "5.2.1",
+    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-react": "7.6.0",
+    "eslint-plugin-standard": "3.0.1",
+    "node-sass": "4.7.2",
+    "nunjucks": "~3",
+    "parcel-bundler": "^1.9.3",
+    "parcel-plugin-nunjucks": "git+https://github.com/The-Politico/parcel-plugin-nunjucks.git",
+    "worker-farm": "^1.6.0"
   }
 }
